product backlog
Product Backlog is an ordered list of everything that is known to be needed in the product. It is the single source of requirements for any changes to be made to the product. The Product Owner is responsible for the Product Backlog, including its content, availability, and ordering. A Product Backlog is never complete. So we tried to sort out our project with all our efforts.
Done by Ashcik Areafin , Yeamin Alam , Abdullah al Nayeem Khan & Saiful Islam .